[#231] refactor : 컴포넌트화를 통한 코드 중복 제거
