+++ src/components/auth/index.ts
@@ -1,0 +1,2 @@
+export { default as MascotCharacter } from "./MascotCharacter";
+export { default as SocialLoginButtons } from "./SocialLoginButtons";
