+++ src/components/auth/signup/SignupForm.tsx
@@ -1,0 +1,170 @@
+"use client";
+
+import React from "react";
+import { FormProvider, useForm } from "react-hook-form";
+import { BaseInput } from "@/components/common/input/BaseInput";
+import { PasswordInput } from "@/components/common/input/PasswordInput";
+import { ISignUpFormValues } from "@/types/auth";
+import { useAuth } from "@/providers/AuthProvider";
+import { useModal } from "@/components/common/modal/ModalContext";
+import { useValidationRules } from "@/hooks/useValidationRules";
+import { useLocale, useTranslations } from "next-intl";
+
+interface SignupFormProps {
+  userType: "CUSTOMER" | "MOVER";
+  signinLink: string;
+}
+
+const SignupForm = ({ userType, signinLink }: SignupFormProps) => {
+  const { signUp, isLoading } = useAuth();
+  const validationRules = useValidationRules();
+  const { open, close } = useModal();
+  const t = useTranslations("auth");
+  const currentLocale = useLocale();
+
+  const form = useForm<ISignUpFormValues>({
+    mode: "onChange",
+  });
+
+  const {
+    register,
+    handleSubmit,
+    watch,
+    formState: { errors, isValid },
+  } = form;
+
+  // 입력 값 감시
+  const name = watch("name");
+  const email = watch("email");
+  const phoneNumber = watch("phoneNumber");
+  const password = watch("password");
+  const passwordCheck = watch("passwordCheck");
+
+  // 회원 가입 버튼 활성화 조건 (값 존재 + validation 통과)
+  const isFormValid = name && email && phoneNumber && password && passwordCheck && isValid;
+
+  const onSubmit = async (data: ISignUpFormValues) => {
+    const signUpData = {
+      email: data.email,
+      password: data.password,
+      name: data.name,
+      phoneNumber: data.phoneNumber,
+      userType: userType as const,
+    };
+
+    try {
+      if (isLoading) return;
+      const response = await signUp(signUpData);
+      if (!response.success) {
+        open({
+          title: "회원가입 실패",
+          children: <div>{response.message}</div>,
+          buttons: [{ text: "확인", onClick: () => close() }],
+        });
+      }
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  return (
+    <FormProvider {...form}>
+      <form onSubmit={handleSubmit(onSubmit)} className="flex w-full flex-col">
+        {/* 이름 */}
+        <div className="mb-6 flex flex-col gap-2 font-normal">
+          <span className="text-black-400 text-md">{t("name")}</span>
+          <BaseInput
+            {...register("name", validationRules.name)}
+            error={errors.name?.message}
+            placeholder={t("namePlaceholder")}
+            inputClassName="py-3.5 px-3.5"
+            wrapperClassName="w-full sm:w-full"
+          />
+        </div>
+
+        {/* 이메일 */}
+        <div className="mb-6 flex flex-col gap-2 font-normal">
+          <span className="text-black-400 text-md">{t("email")}</span>
+          <BaseInput
+            {...register("email", validationRules.email)}
+            error={errors.email?.message}
+            placeholder={t("emailPlaceholder")}
+            inputClassName="py-3.5 px-3.5"
+            wrapperClassName="w-full sm:w-full"
+          />
+        </div>
+
+        {/* 전화번호 */}
+        <div className="mb-6 flex flex-col gap-2 font-normal">
+          <span className="text-black-400 text-md">{t("phoneNumber")}</span>
+          <BaseInput
+            {...register("phoneNumber", validationRules.phoneNumber)}
+            error={errors.phoneNumber?.message}
+            placeholder={t("phoneNumberPlaceholder")}
+            inputClassName="py-3.5 px-3.5"
+            wrapperClassName="w-full sm:w-full"
+          />
+        </div>
+
+        {/* 비밀번호*/}
+        <div className={`flex flex-col gap-2 ${currentLocale === "en" ? "mb-12" : "mb-6"}`}>
+          <span className="text-black-400 text-md font-normal">{t("password")}</span>
+          <PasswordInput
+            {...register("password", validationRules.password)}
+            placeholder={t("passwordPlaceholder")}
+            inputClassName="py-3.5 px-3.5"
+            wrapperClassName="w-full sm:w-full"
+          />
+        </div>
+
+        {/* 비밀번호 확인*/}
+        <div className="mb-6 flex flex-col gap-2">
+          <span className="text-black-400 text-md font-normal">{t("passwordCheck")}</span>
+          <PasswordInput
+            {...register("passwordCheck", {
+              required: t("passwordCheckRequired"),
+              validate: (value) => {
+                const passwordValue = watch("password");
+                return value === passwordValue || t("passwordMismatch");
+              },
+            })}
+            placeholder={t("passwordCheckPlaceholder")}
+            inputClassName="py-3.5 px-3.5"
+            wrapperClassName="w-full sm:w-full"
+          />
+        </div>
+
+        <div className="flex flex-col gap-4">
+          {/* 회원가입 버튼 */}
+          <button
+            type="submit"
+            disabled={!isFormValid || isLoading}
+            className={`mt-4 rounded-xl px-4 py-4 text-lg font-semibold text-white transition ${
+              isFormValid && !isLoading
+                ? "bg-primary-400 hover:bg-primary-500 cursor-pointer"
+                : "cursor-not-allowed bg-gray-300"
+            }`}
+          >
+            {isLoading ? (
+              <div className="flex items-center justify-center">
+                <div className="h-4 w-4 animate-spin rounded-full border-2 border-white border-t-transparent" />
+                <span className="ml-2">{t("signupInProgress")}</span>
+              </div>
+            ) : (
+              t("signup")
+            )}
+          </button>
+          {/* 로그인 링크 */}
+          <div className="flex w-full items-center justify-center gap-2">
+            <span className="text-black-200 text-lg">{t("alreadyMember")}</span>
+            <a href={signinLink}>
+              <span className="text-primary-400 text-lg font-semibold underline">{t("login")}</span>
+            </a>
+          </div>
+        </div>
+      </form>
+    </FormProvider>
+  );
+};
+
+export default SignupForm;
