+++ src/components/auth/signup/index.ts
@@ -1,0 +1,2 @@
+export { default as SignupForm } from "./SignupForm";
+export { default as SignupHeader } from "./SignupHeader";
